Multi-carton list creation
Under the multicarton_entry folder, I've begun working on a method for the user to copy and paste a body of text (from the manifest) and it will sort through line by line only taking lines that have the desired vendor for this project. At some point, we can add other vendors to the list or simply have it be "all vendors".